move test_FBS to new script
